Split tracking models into separate files
We're starting to add more business logic to the tracking models as well as
accumulate more of them, which makes keeping them all in one file a little
cumbersome. Split them into individual files.

No changes to the actual code here.
