+++ src/main/kotlin/com/terraformation/backend/tracking/model/Constants.kt
@@ -1,0 +1,4 @@
+package com.terraformation.backend.tracking.model
+
+/** Number of square meters in a hectare. */
+const val SQUARE_METERS_PER_HECTARE: Double = 10000.0
