+++ src/main/kotlin/com/terraformation/backend/tracking/model/DeliveryModel.kt
@@ -1,0 +1,25 @@
+package com.terraformation.backend.tracking.model
+
+import com.terraformation.backend.db.nursery.WithdrawalId
+import com.terraformation.backend.db.tracking.DeliveryId
+import com.terraformation.backend.db.tracking.PlantingSiteId
+import com.terraformation.backend.db.tracking.tables.references.DELIVERIES
+import org.jooq.Field
+import org.jooq.Record
+
+data class DeliveryModel(
+    val id: DeliveryId,
+    val plantings: List<PlantingModel>,
+    val plantingSiteId: PlantingSiteId,
+    val withdrawalId: WithdrawalId,
+) {
+  constructor(
+      record: Record,
+      plantingsMultisetField: Field<List<PlantingModel>>
+  ) : this(
+      record[DELIVERIES.ID]!!,
+      record[plantingsMultisetField],
+      record[DELIVERIES.PLANTING_SITE_ID]!!,
+      record[DELIVERIES.WITHDRAWAL_ID]!!,
+  )
+}
