+++ src/main/kotlin/com/terraformation/backend/tracking/model/PlantingSiteDepth.kt
@@ -1,0 +1,8 @@
+package com.terraformation.backend.tracking.model
+
+enum class PlantingSiteDepth {
+  Site,
+  Zone,
+  Subzone,
+  Plot
+}
