+++ src/main/kotlin/com/terraformation/backend/tracking/model/PlantingZoneModel.kt
@@ -1,64 +1,11 @@
 package com.terraformation.backend.tracking.model
 
-import com.terraformation.backend.db.default_schema.OrganizationId
-import com.terraformation.backend.db.default_schema.SpeciesId
-import com.terraformation.backend.db.nursery.WithdrawalId
-import com.terraformation.backend.db.tracking.DeliveryId
 import com.terraformation.backend.db.tracking.MonitoringPlotId
-import com.terraformation.backend.db.tracking.PlantingId
-import com.terraformation.backend.db.tracking.PlantingSiteId
 import com.terraformation.backend.db.tracking.PlantingSubzoneId
-import com.terraformation.backend.db.tracking.PlantingType
 import com.terraformation.backend.db.tracking.PlantingZoneId
-import com.terraformation.backend.db.tracking.tables.references.DELIVERIES
-import com.terraformation.backend.db.tracking.tables.references.PLANTINGS
-import com.terraformation.backend.db.tracking.tables.references.PLANTING_SITES
 import com.terraformation.backend.util.equalsIgnoreScale
 import java.math.BigDecimal
-import java.time.Month
-import java.time.ZoneId
-import org.jooq.Field
-import org.jooq.Record
 import org.locationtech.jts.geom.MultiPolygon
-import org.locationtech.jts.geom.Polygon
-
-data class MonitoringPlotModel(
-    val boundary: Polygon,
-    val id: MonitoringPlotId,
-    val fullName: String,
-    val name: String,
-    val permanentCluster: Int? = null,
-    val permanentClusterSubplot: Int? = null,
-) {
-  fun equals(other: Any?, tolerance: Double): Boolean {
-    return other is MonitoringPlotModel &&
-        id == other.id &&
-        fullName == other.fullName &&
-        name == other.name &&
-        permanentCluster == other.permanentCluster &&
-        permanentClusterSubplot == other.permanentClusterSubplot &&
-        boundary.equalsExact(other.boundary, tolerance)
-  }
-}
-
-data class PlantingSubzoneModel(
-    val areaHa: BigDecimal,
-    val boundary: MultiPolygon,
-    val id: PlantingSubzoneId,
-    val fullName: String,
-    val name: String,
-    val monitoringPlots: List<MonitoringPlotModel>,
-) {
-  fun equals(other: Any?, tolerance: Double): Boolean {
-    return other is PlantingSubzoneModel &&
-        id == other.id &&
-        fullName == other.fullName &&
-        name == other.name &&
-        areaHa.equalsIgnoreScale(other.areaHa) &&
-        monitoringPlots.zip(other.monitoringPlots).all { it.first.equals(it.second, tolerance) } &&
-        boundary.equalsExact(other.boundary, tolerance)
-  }
-}
 
 data class PlantingZoneModel(
     val areaHa: BigDecimal,
@@ -156,96 +103,3 @@ data class PlantingZoneModel(
         boundary.equalsExact(other.boundary, tolerance)
   }
 }
-
-data class PlantingSiteModel(
-    val areaHa: BigDecimal? = null,
-    val boundary: MultiPolygon?,
-    val description: String?,
-    val id: PlantingSiteId,
-    val name: String,
-    val organizationId: OrganizationId,
-    val plantingSeasonEndMonth: Month? = null,
-    val plantingSeasonStartMonth: Month? = null,
-    val plantingZones: List<PlantingZoneModel>,
-    val timeZone: ZoneId? = null,
-) {
-  constructor(
-      record: Record,
-      plantingZonesMultiset: Field<List<PlantingZoneModel>>? = null
-  ) : this(
-      areaHa = record[PLANTING_SITES.AREA_HA],
-      boundary = record[PLANTING_SITES.BOUNDARY] as? MultiPolygon,
-      description = record[PLANTING_SITES.DESCRIPTION],
-      id = record[PLANTING_SITES.ID]!!,
-      name = record[PLANTING_SITES.NAME]!!,
-      organizationId = record[PLANTING_SITES.ORGANIZATION_ID]!!,
-      plantingSeasonEndMonth = record[PLANTING_SITES.PLANTING_SEASON_END_MONTH],
-      plantingSeasonStartMonth = record[PLANTING_SITES.PLANTING_SEASON_START_MONTH],
-      plantingZones = plantingZonesMultiset?.let { record[it] } ?: emptyList(),
-      timeZone = record[PLANTING_SITES.TIME_ZONE],
-  )
-
-  fun equals(other: Any?, tolerance: Double): Boolean {
-    return other is PlantingSiteModel &&
-        description == other.description &&
-        id == other.id &&
-        name == other.name &&
-        timeZone == other.timeZone &&
-        plantingSeasonEndMonth == other.plantingSeasonEndMonth &&
-        plantingSeasonStartMonth == other.plantingSeasonStartMonth &&
-        plantingZones.size == other.plantingZones.size &&
-        areaHa.equalsIgnoreScale(other.areaHa) &&
-        plantingZones.zip(other.plantingZones).all { it.first.equals(it.second, tolerance) } &&
-        (boundary == null && other.boundary == null ||
-            boundary != null &&
-                other.boundary != null &&
-                boundary.equalsExact(other.boundary, tolerance))
-  }
-}
-
-data class PlantingModel(
-    val id: PlantingId,
-    val notes: String? = null,
-    val numPlants: Int,
-    val plantingSubzoneId: PlantingSubzoneId? = null,
-    val speciesId: SpeciesId,
-    val type: PlantingType,
-) {
-  constructor(
-      record: Record
-  ) : this(
-      record[PLANTINGS.ID]!!,
-      record[PLANTINGS.NOTES],
-      record[PLANTINGS.NUM_PLANTS]!!,
-      record[PLANTINGS.PLANTING_SUBZONE_ID],
-      record[PLANTINGS.SPECIES_ID]!!,
-      record[PLANTINGS.PLANTING_TYPE_ID]!!,
-  )
-}
-
-data class DeliveryModel(
-    val id: DeliveryId,
-    val plantings: List<PlantingModel>,
-    val plantingSiteId: PlantingSiteId,
-    val withdrawalId: WithdrawalId,
-) {
-  constructor(
-      record: Record,
-      plantingsMultisetField: Field<List<PlantingModel>>
-  ) : this(
-      record[DELIVERIES.ID]!!,
-      record[plantingsMultisetField],
-      record[DELIVERIES.PLANTING_SITE_ID]!!,
-      record[DELIVERIES.WITHDRAWAL_ID]!!,
-  )
-}
-
-enum class PlantingSiteDepth {
-  Site,
-  Zone,
-  Subzone,
-  Plot
-}
-
-/** Number of square meters in a hectare. */
-const val SQUARE_METERS_PER_HECTARE: Double = 10000.0
